I will now implement function-based initialization.
This commit introduces a new function-based initialization system to improve the stability and reliability of the addon's setup process.

I will make the following changes:
- Add `AntiCheatsBP/functions/ac.mcfunction` to trigger initialization.
- Create `AntiCheatsBP/scripts/core/initializationManager.js` to handle the initialization logic.
- Refactor `AntiCheatsBP/scripts/main.js` to use the new initialization system.
- Update `README.md` with the new setup instructions.

After these changes, you will need to run `/function ac` to initialize the addon for the first time.
