+++ AntiCheatsBP/functions/ac.mcfunction
@@ -1,0 +1,1 @@
+scriptevent ac:init
